Factor out function to add `noqa` comments to code
This reduces duplication between the `flake8` and `ruff` plugins.
